Consolidate logic and ensure we always paginate through result sets
BigQuery some times says a job is complete even when it's only returned
the first row. We were trying to handle these flows in several different
places, but, as is the case when logic is duplicated, we missed some spots.

This commit DRYs up some of that duplication and ensures that we always
iterate against the full result set -- independent of the "complete" flag
returned by BigQuery.