Retry if job reference is null
For whatever reason, periodically we do not get a job reference when
queuing a query, but we also don't get an error. In these cases, retry
a few times before giving up.

Additionally, if we never get a job reference, we have a hard stop.
Otherwise the script starts to panic.